refactor: break esplora api types into new file
