+++ core/src/prelude.rs
@@ -1,9 +1,6 @@
 //! Re-exports of common traits.
 pub use crate::{
-    builder::TxBuilder,
-    enc::AddressEncoder,
-    nets::Network,
-    hashes::marked::MarkedDigest,
+    builder::TxBuilder, enc::AddressEncoder, hashes::marked::MarkedDigest, nets::Network,
     ser::ByteFormat, types::Transaction,
 };
 pub use bitcoin_spv::types::Hash256Digest;
